Data Pipeline Refactor: Move Tabular to new DataModule (#1049)
